adding test for adapting the mesh
